Avoid channel name and topic edits
Discord changed their API to severely rate-limit changes to channel names and topics, to 2 per 10 minutes. This has significant ramifications on how regions can be managed.

In order to minimize this effect, a region's active status is not maintained in the description - thus, the name and description only need to be touched if an admin manually changes them or forces an update.